+++ packages/toolbox/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @interunit/toolbox
 
+## 0.0.14
+
+### Patch Changes
+
+- - Combobox
+    - Alpha version ready for testing
+    - Tests
+    - Initial Docs
+  - Popover
+    - Make sure Props are passing through Popover Elements
+    - Calculate alignment based on content and not just the trigger height (may need to adjust this some more
+  - Toolbox
+    - Add dispatch type as optional union type for onValueChange
+  - Modal
+    - Adjust ModalProps to use intrinsic elements instead of HTMLProps (Modal generally needs more attention)
+
 ## 0.0.13
 
 ### Patch Changes